Support universal fallback config
Instead of requiring every filetype to have a cmd specified, support a
fallback so you can still open a repl when editing readmes or other
non-core filetype files. Better to open your fallback (likely the shell)
than give an error. This allows mappings that open or focus the repl
because without this change that mapping needs to determine if the repl
is already open so they can pass the cmd.

This is preferable to changing zepl#start to use zepl#config('cmd',
&shell) since users might need other configuration options for their
cmd. Downside is that the config doesn't change if user's 'shell'
changes, but that seems unlikely/infrequent.

Our config are entirely self-contained -- there's no individual key
fallbacks so there's no bleed over from FALLBACK into other configs.

Test
* open .txt, gZ (shell repl), open .py, gZ (back to shell), close repl,
  gZ (python repl).
    nnoremap gZ <Cmd>Repl<CR>
    let g:repl_config = {
                \     'FALLBACK': {
                \         'cmd': &shell,
                \     },
                \     'python': {
                \         'cmd': 'python',
                \     },
                \ }